backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
reportlab==4.2.5
python-dateutil==2.9.0
pytz==2024.2
httpx==0.27.0
